// citpwawcl/src/app/models/dati-cg-model.ts
export class DatiCGModel {
    idTipoComponente: string;
    progressivo: number;
    dataInstall: string;
    codiceImpianto: number;
    dataDismiss: string;
    flgDismissione: number;
    dataUltMod: string;
    utenteUltMod: string;
    fkMarca: number;
    descMarca: string;
    fkCombustibile: number;
    descCombustibile: string;
    matricola: string;
    modello: string;
    potenzaTermicaKw: number;
    note: string;
    tempoManutAnni: number;
    dataMinimaControllo: string;
    dataMassimaControllo: string;
    tipologia: string;
    potenzaElettricaKw: number;
    tempH2oOutMin: number;
    tempH2oOutMax: number;
    tempH2oInMin: number;
    tempH2oInMax: number;
    tempH2oMotoreMin: number;
    tempH2oMotoreMax: number;
    tempFumiValleMin: number;
    tempFumiValleMax: number;
    tempFumiMonteMin: number;
    tempFumiMonteMax: number;
    coMin: number;
    coMax: number;
    alimentazione: string;
    cf: string;
    siglaRea: string;
    numeroRea: string;
}